Fixed VectorSpace::operator-, split logic from Cluster.hpp into Cluster.cpp, add standart template in KMeans
