+++ include/Cluster.hpp
@@ -23,20 +23,6 @@ private:
     std::vector<size_t> indexes;
 };
 
-Cluster::Cluster() {}
-
-Cluster::Cluster(const std::vector<size_t> &indexes) : indexes(indexes) {}
-
-void Cluster::addIndexElem(size_t index)
-{
-    indexes.push_back(index);
-}
-
-std::vector<size_t> Cluster::getClusteringIndexes() const
-{
-    return indexes;
-}
-
 template <typename Data>
 std::vector<Data> Cluster::getClusteringDataByData(const std::vector<Data> &data) const
 {
