+++ include/VectorSpace.hpp
@@ -84,7 +84,7 @@ double VectorSpace<T>::operator-(const VectorSpace<T> &obj) const
     size_t i = 0;
     for (const T &elem : vec)
     {
-        ans += std::abs(elem - obj.vec[i]);
+        ans += std::abs(elem - obj.vec[i++]);
     }
     return ans;
 }
